통화 변환의 타입을 금액이 아니라 통화가 정하게
테스트가 실제 버그를 잡았다. PHP 의 / 는 나누어떨어지면 int 를
돌려준다.

    USD 9900 → 99   (int)    JSON 에 99
    USD 9950 → 99.5 (float)  JSON 에 99.5

같은 통화인데 금액에 따라 JSON 모양이 달라진다. 받는 쪽이 타입을
보고 분기하면 특정 금액에서만 어긋난다. 소수 자릿수가 있는 통화는
언제나 float 로 내보낸다.

USD 9950 케이스를 테스트에 추가했다.

+++ src/Channel/Ga4Channel.php
@@ -165,7 +165,12 @@ final class Ga4Channel implements ChannelInterface
             default => 2,
         };
 
-        return $exponent === 0 ? $minor : $minor / (10 ** $exponent);
+        // PHP 의 / 는 나누어떨어지면 int 를 돌려준다.
+        //   USD 9900 → 99   (int)
+        //   USD 9950 → 99.5 (float)
+        // 같은 통화인데 금액에 따라 JSON 모양이 달라진다. 타입은
+        // 금액이 아니라 **통화**가 정해야 하므로 명시적으로 캐스팅한다.
+        return $exponent === 0 ? $minor : (float) ($minor / (10 ** $exponent));
     }
 
     /**
